Added in build for common library
The common library, mostly taken from the NVIDIA CUDA SDK contains a wealth of
various utility functions that can be used to perform aspects such as
benchmarking kernels, asserting results debugging etc.
We will probably add to the library over time, but for now we aim to get it
linked into test binaries as part of the build.